python-backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
aniso8601==10.0.0
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
attrs==25.1.0
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
eventlet==0.39.0
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
flask-restx==1.3.0
Flask-SocketIO==5.5.1
frozenlist==1.5.0
fsspec==2025.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
langchain==0.3.17
langchain-anthropic==0.3.5
langchain-community==0.3.16
langchain-core==0.3.33
langchain-google-genai==2.0.9
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langsmith==0.3.4
MarkupSafe==3.0.2
marshmallow==3.26.0
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
nanoid==2.0.0
neo4j==5.27.0
numpy==1.26.4
openai==1.61.0
orjson==3.10.15
packaging==24.2
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pymongo==4.11
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.48.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
